+++ lib/zaikio/procurement.rb
@@ -22,6 +22,8 @@ require "zaikio/procurement/delivery"
 require "zaikio/procurement/delivery_line_item"
 require "zaikio/procurement/current_supplier"
 require "zaikio/procurement/material_requirement"
+require "zaikio/procurement/material_availability_check"
+
 module Zaikio
   module Procurement
     class << self