+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Updated MaterialRequirement attributes
+
 ## [0.1.0] - 2021-10-05
 
 * Added support for material requirements