+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Added MaterialAvailabilityCheck for suppliers
+
 ## [0.1.1] - 2021-10-06
 
 ### Changed
@@ -14,5 +19,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added support for material requirements
 
-[Unreleased]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.1.0..HEAD
+[Unreleased]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.1.1..HEAD
+[0.1.1]: https://github.com/zaikio/zaikio-procurement-ruby/compare/v0.1.0..v0.1.1
 [0.1.0]: https://github.com/zaikio/zaikio-procurement-ruby/compare/3e84659a2eee172280a7e4f0434fd8ce0e373844..8c9a509b308290ba18c17ef68701fd451cb05d18